raw-parse/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
pandas
google-cloud-bigquery